Do not treat warnings as errors
Appveyor build fails treating warnings as errors. The warning is that
License is non-standard. The license issue was debated when the
package was created for @ropensci:

https://github.com/ropensci/software-review/issues/118#issuecomment-317260241

Message:
> checking DESCRIPTION meta-information ... WARNING
> Non-standard license specification:
>   CC0-1.0 + file LICENSE

https://ci.appveyor.com/project/timtrice/rrricanesdata-on6xt/builds/21296022/job/n3iygk4wgn1stu00#L276

Source:
https://github.com/ropensci/rrricanesdata/blob/master/appveyor.yml#L16

Fix:
The issue lies in the WARNINGS_ARE_ERRORS. Per the issue below (and the
latest README for kmrlmlr/r-appveyor), set this to FALSE by leaving
the field empty.

https://github.com/krlmlr/r-appveyor/issues/79
https://github.com/krlmlr/r-appveyor/blob/ac6254e6d2e8629e037a0ddcfde985661487cd0a/README.md